feat: Creacion de recordatorios automaticos

+++ src/main/java/com/petwellness/repository/RecordatorioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecordatorioRepository extends JpaRepository<Recordatorio, Long> {
+public interface RecordatorioRepository extends JpaRepository<Recordatorio, Integer> {
 }
